Move `/` and `/preferences` routes to seperate files
