+++ src/app.js
@@ -3,7 +3,8 @@ const helmet = require("helmet");
 const routes = express.Router();
 
 const news = require("./routes/news.route");
-const { login, register } = require("./controllers/auth.controller");
+const preferences = require("./routes/preferences.route");
+const root = require("./routes/root.route");
 
 let PORT;
 const app = express();
@@ -13,18 +14,9 @@ app.use(routes);
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 
-routes.use(express.json());
-routes.use(express.urlencoded({ extended: false }));
-
-routes.get("/", (_, res) => {
-  res.status(200).send("<h3>Welcome to the News Aggregator API</h3>");
-});
-
+routes.use("/", root);
 routes.use("/news", news);
-
-routes.post("/register", register);
-
-routes.post("/login", login);
+routes.use("/preferences", preferences);
 
 if (process.env.PORT !== "") {
   PORT = process.env.PORT;
