+++ src/users.json
@@ -1,1 +1,1 @@
-[{"id":"101","username":"chinmay","password":"$2b$08$ToXC0n.hqXvPXHLM5.x4nedZdP1NN24zjY6b9CMkrP31e75PXqbNq","preferences":{"sources":["bbc-news","cnn"],"categories":["business","technology"]}}]
+[{"id":"101","username":"chinmay","password":"$2b$08$ToXC0n.hqXvPXHLM5.x4nedZdP1NN24zjY6b9CMkrP31e75PXqbNq","preferences":{"sources":["google-news-in"],"categories":["health","technology","sports"]}}]
